requirements.txt
discord.py
python-dotenv
better_profanity